fix - debug

+++ utilities/install_runtime_prerequisites.sh
@@ -14,17 +14,33 @@ dpkg --configure -a
 #
 # * Basic
 apt-get -y install ssh pdsh curl expect libxml2-utils
-# * Oracle JDK 1.8 (preferred)
-apt-get -y purge openjdk*
-add-apt-repository -y ppa:webupd8team/java
-apt-get -y update
-echo oracle-java8-installer shared/accepted-oracle-license-v1-1 select true | /usr/bin/debconf-set-selections
-apt-get -y install oracle-java8-installer --allow-unauthenticated
-apt-get -y install oracle-java8-set-default --allow-unauthenticated
-apt-get -y install software-properties-common
 
-if [[ '' == `which java` ]]; then 
-    apt-get -y install openjdk-8-jdk; 
+# FOR UBUNTU
+cat /etc/*release | grep ID=ubuntu
+if [[ $? == 0 ]]; then
+	# * Oracle JDK 1.8 (preferred)
+	apt-get -y purge openjdk*
+	add-apt-repository -y ppa:webupd8team/java
+	apt-get -y update
+	echo oracle-java8-installer shared/accepted-oracle-license-v1-1 select true | /usr/bin/debconf-set-selections
+	apt-get -y install oracle-java8-installer --allow-unauthenticated
+	apt-get -y install oracle-java8-set-default --allow-unauthenticated
+	apt-get -y install software-properties-common
+
+	# * Open JDK 1.8
+	if [[ '' == `which java` ]]; then 
+	    apt-get -y install openjdk-8-jdk; 
+	fi
+fi
+
+# FOR DEBIAN
+cat /etc/*release | grep ID=debian
+if [[ $? == 0 ]]; then
+	apt-get -f install apt-transport-https
+	echo deb http://http.debian.net/debian jessie-backports main >> /etc/apt/sources.list
+	apt-get -f update
+	apt-get -f install openjdk-8-jdk
+	# update-alternatives --config java
 fi
 
 # ----------------------------
